+++ test/rule-handling-methods-test.js
@@ -440,34 +440,13 @@ describe('Tokenizer RuleSet Methods', function () {
     })
   })
 
-  // Check the existence of a rule
-  describe('#existsRule', function () {
-    describe('does exist', function () {
-      var p = new Tokenizer(options)
-      it('should return true', function (done) {
-        p.addRule('a', 'first')
-        assert.equal(p.existsRule('first'), true)
-        done()
-      })
-    })
-
-    describe('does not exist', function () {
-      var p = new Tokenizer(options)
-      it('should return false', function (done) {
-        p.addRule('a', 'first')
-        assert.equal(p.existsRule('second'), false)
-        done()
-      })
-    })
-  })
-
   // Delete a rule set
-  describe('#deleteRuleSet', function () {
+  describe('#removeRuleSet', function () {
     var p = new Tokenizer(options)
     it('should remove it', function (done) {
       p.addRule('a', 'first')
       p.saveRuleSet('myRules')
-      p.deleteRuleSet('myRules')
+      p.removeRuleSet('myRules')
       assert.throws(
         function () {
           p.loadRuleSet('myRules')
@@ -476,19 +455,6 @@ describe('Tokenizer RuleSet Methods', function () {
           if (err instanceof Error) return true
         }
       )
-      done()
-    })
-  })
-
-  // Get all rule sets
-  describe('#getAllRuleSet', function () {
-    var p = new Tokenizer(options)
-    it('should return rule sets: myRules and myRules2', function (done) {
-      p.addRule('a', 'first')
-      p.saveRuleSet('myRules')
-      p.addRule('b', 'second')
-      p.saveRuleSet('myRules2')
-      assert.deepEqual(Object.keys( p.getAllRuleSet() ), ['myRules','myRules2'])
       done()
     })
   })